+++ test/case-diff.sh
@@ -1,3 +1,5 @@
 
-./xsvutils -v2 ./test/data/sample-4.tsv diff ./test/data/sample-5.tsv > ./var/test/diff.txt
+./xsvutils ./test/data/sample-4.tsv diff    ./test/data/sample-5.tsv > ./var/test/diff-1.txt
+./xsvutils ./test/data/sample-4.tsv diff -b ./test/data/sample-5.tsv > ./var/test/diff-2.txt
+./xsvutils ./test/data/sample-4.tsv diff -w ./test/data/sample-5.tsv > ./var/test/diff-3.txt