+++ CHANGELOG.md
@@ -6,6 +6,7 @@ To see compatibility information, `xsvutils help compatibility`.
 
 ## version 0.3.4 (Unreleased)
 - add `ratio` subcommand
+- add `-b`, `-w` options of `diff` subcommand
 
 ## version 0.3.3 (2018/05/27)
 - add `rmnoname` subcommand