+++ CHANGELOG.md
@@ -4,7 +4,7 @@ To see version information, `xsvutils --version`.
 
 To see compatibility information, `xsvutils help compatibility`.
 
-## version 0.3.4 (Unreleased)
+## version 0.3.4 (2018/06/30)
 - add `ratio` subcommand
 - add `-b`, `-w` options of `diff` subcommand
 - add `--o-chart` option as output format